% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbal_oil_forecast.R
\name{mbal_forecast_oil.volumetric_forecast_oil}
\alias{mbal_forecast_oil.volumetric_forecast_oil}
\title{S3 method for class 'mbal_forecast_oil'}
\usage{
\method{mbal_forecast_oil}{volumetric_forecast_oil}(forecast_lst, time_lst)
}
\arguments{
\item{forecast_lst}{a list object of class 'forecast_oil'}

\item{time_lst}{a list object of class 'time'}
}
\value{
a data frame with estimates for saturation of fluids, liquid dropout, gas-oil ratio, recovery factor, and drive indices over a range of given pressures for a volumetric oil reservoir
}
\description{
Return a data frame with estimates for saturation of fluids, liquid dropout, gas-oil ratio, recovery factor, and drive indices over a range of given pressures for a volumetric oil reservoir
}
