% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbal_gas_optim.R
\name{mbal_optim_gas.water_drive_optim_gas}
\alias{mbal_optim_gas.water_drive_optim_gas}
\title{S3 method for class 'mbal_optim_gas'}
\usage{
\method{mbal_optim_gas}{water_drive_optim_gas}(optim_lst, time_lst)
}
\arguments{
\item{optim_lst}{a list object of class 'optimization_gas'}

\item{time_lst}{a list object of class 'time'}
}
\value{
a list of class 'mbal_gas' with estimates for the unknown parameters of a water_drive gas reservoir
}
\description{
Generate a list of class 'mbal_gas' with estimates for the unknown parameters of a water_drive gas reservoir
}
