% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbal_gas_model.R
\name{mbal_perform_gas.volumetric_gas}
\alias{mbal_perform_gas.volumetric_gas}
\title{S3 method for class 'mbal_perform_gas'}
\usage{
\method{mbal_perform_gas}{volumetric_gas}(mbal_lst, time_lst)
}
\arguments{
\item{mbal_lst}{a list object of class 'mbal_gas'}

\item{time_lst}{a list object of class 'time'}
}
\value{
a data frame with estimates for fluids saturation, drive indices, production rates, and gas-oil ratios over the pressure history of a volumetric gas reservoir
}
\description{
Return a data frame with estimates for fluids saturation, drive indices, production rates, and gas-oil ratios over the pressure history of a volumetric gas reservoir
}
