% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbal_oil_model.R
\name{mbal_perform_oil.gas_cap_oil}
\alias{mbal_perform_oil.gas_cap_oil}
\title{S3 method for class 'mbal_perform_oil'}
\usage{
\method{mbal_perform_oil}{gas_cap_oil}(mbal_lst, time_lst)
}
\arguments{
\item{mbal_lst}{a list object of class 'mbal_oil'}

\item{time_lst}{a list object of class 'time'}
}
\value{
a data frame with estimates for fluids saturation, drive indices, production rates, and gas-oil ratios over the pressure history of a gas_cap_drive oil reservoir
}
\description{
Return a data frame with estimates for fluids saturation, drive indices, production rates, and gas-oil ratios over the pressure history of a gas_cap_drive oil reservoir
}
